# specs/002-fastapi-dashboard-api/contracts/openapi-spec.yaml
openapi: 3.0.3
info:
  title: RiseTrader Dashboard API
  description: |
    FastAPI service serving market data, forecasts, account information, and strategy allocations
    from PostgreSQL database to React dashboard with real-time MT4 integration.

    ## Real-Time Updates
    The API provides Server-Sent Events (SSE) endpoints for real-time data streaming:
    - `/api/stream/market-data` - Price updates
    - `/api/stream/account` - Account balance and positions
    - `/api/stream/forecasts` - New forecast availability

    ## Authentication
    ⚠️ Currently NO authentication implemented. API key + JWT authentication will be added in Phase 8-9.

    ## Rate Limiting
    Rate limiting is enforced via slowapi middleware. Default limits apply per IP address.

    ## Pagination
    List endpoints use keyset (cursor-based) pagination for consistent performance:
    - `page`: Page number (1-indexed)
    - `page_size`: Items per page (default: 50, max: 500)
    - `cursor`: Opaque cursor for next page (returned in `next_cursor` field)
  version: 1.0.0
  contact:
    name: RiseTrader Team
    email: support@risetrader.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8003/api
    description: Local development server
  - url: https://api.risetrader.com/api
    description: Production server (future)
...
paths:
...
  /market-data/{symbol}:
    get:
      tags:
        - market-data
      summary: Get market data for symbol
      description: |
        Retrieve latest market data ticks for a specific trading symbol.
        Supports filtering by timeframe and pagination.

        **Performance**: <2 seconds for 500 candlesticks (success criteria SC-002)
      operationId: getMarketData
      parameters:
        - name: symbol
          in: path
          required: true
          description: Trading symbol (e.g., CrudeOIL, EURUSD)
          schema:
            type: string
            example: CrudeOIL
        - name: timeframe
          in: query
          required: false
          description: Timeframe filter
          schema:
            type: string
            enum: [M1, M5, M15, M30, H1, H4, D1, W1, MN1]
            example: M5
        - name: limit
          in: query
          required: false
          description: Maximum number of records to return
          schema:
            type: integer
            default: 500
            minimum: 1
            maximum: 1000
        - name: page
          in: query
          required: false
          description: Page number (1-indexed)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          required: false
          description: Items per page
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 500
        - name: cursor
          in: query
          required: false
          description: Cursor for pagination (keyset pagination)
          schema:
            type: string
            example: "2024-11-24T10:30:00Z"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketDataListResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Symbol not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /market-data/{symbol}/range:
    get:
      tags:
        - market-data
      summary: Get market data for time range
      description: Retrieve market data for a specific time range
      operationId: getMarketDataRange
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: start_time
          in: query
          required: true
          description: Start timestamp (ISO 8601)
          schema:
            type: string
            format: date-time
            example: "2024-11-24T00:00:00Z"
        - name: end_time
          in: query
          required: true
          description: End timestamp (ISO 8601)
          schema:
            type: string
            format: date-time
            example: "2024-11-24T23:59:59Z"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketDataListResponse'
        '400':
          description: Invalid date range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    MarketDataResponse:
      type: object
      required:
        - id
        - time
        - symbol
        - timeframe
        - open
        - high
        - low
        - close
        - volume
      properties:
        id:
          type: integer
          example: 12345
        time:
          type: string
          format: date-time
          example: "2024-11-24T10:30:00Z"
        symbol:
          type: string
          example: "CrudeOIL"
        timeframe:
          type: string
          enum: [M1, M5, M15, M30, H1, H4, D1, W1, MN1]
          example: "M5"
        source:
          type: string
          enum: [BARCHART, MT4]
          example: "MT4"
        open:
          type: number
          format: double
          example: 75.25
        high:
          type: number
          format: double
          example: 75.50
        low:
          type: number
          format: double
          example: 75.20
        close:
          type: number
          format: double
          example: 75.45
        volume:
          type: integer
          example: 1523
        change:
          type: number
          format: double
          example: 0.20
        change_percent:
          type: number
          format: double
          example: 0.27
    MarketDataListResponse:
      type: object
      required:
        - data
        - total
        - page
        - page_size
        - symbol
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MarketDataResponse'
        total:
          type: integer
          example: 1523
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 50
        next_cursor:
          type: string
          nullable: true
          example: "2024-11-24T09:00:00Z"
        symbol:
          type: string
          example: "CrudeOIL"
        start_time:
          type: string
          format: date-time
          nullable: true
        end_time:
          type: string
          format: date-time
          nullable: true
...
    ErrorResponse:
      type: object
      required:
        - error
        - timestamp
      properties:
        error:
          type: string
          example: "Symbol not found"
        detail:
          type: string
          nullable: true
          example: "No data available for symbol INVALID"
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          example: "req_1234567890"
...